Some renaming and moving
