+++ openquake/hazardlib/source/point.py
@@ -23,14 +23,15 @@ from openquake.baselib.general import AccumDict, groupby_grid
 from openquake.baselib.performance import Monitor
 from openquake.hazardlib.geo import Point, geodetic
 from openquake.hazardlib.geo.nodalplane import NodalPlane
-from openquake.hazardlib.source.base import (
-    ParametricSeismicSource, build_planar_surfaces, surfin_dt)
+from openquake.hazardlib.geo.surface.planar import (
+    build_planar_surfaces, planin_dt)
+from openquake.hazardlib.source.base import ParametricSeismicSource
 from openquake.hazardlib.source.rupture import (
     ParametricProbabilisticRupture, PointRupture)
 from openquake.hazardlib.geo.utils import get_bounding_box, angular_distance
 
 
-def _get_rupture_dimensions(surfin):
+def _get_rupture_dimensions(planin):
     """
     Calculate and return the rupture length and width
     for given magnitude surface parameters.
@@ -50,14 +51,14 @@ def _get_rupture_dimensions(surfin):
     depth, the rupture width is shrunken to a maximum possible
     and rupture length is extended to preserve the same area.
     """
-    rup_length = math.sqrt(surfin.area * surfin.rar)
-    rup_width = surfin.area / rup_length
-    seismogenic_layer_width = surfin.lsd - surfin.usd
-    rdip = math.radians(surfin.dip)
+    rup_length = math.sqrt(planin.area * planin.rar)
+    rup_width = planin.area / rup_length
+    seismogenic_layer_width = planin.lsd - planin.usd
+    rdip = math.radians(planin.dip)
     max_width = seismogenic_layer_width / math.sin(rdip)
     if rup_width > max_width:
         rup_width = max_width
-        rup_length = surfin.area / rup_width
+        rup_length = planin.area / rup_width
     return numpy.array([rup_length, rup_width * math.cos(rdip),
                         rup_width * math.sin(rdip)])
 
@@ -124,8 +125,8 @@ def _gen_ruptures(src, nplanes=(), hypos=(), shift_hypo=False, step=1):
         clon, clat, cdeps = hypo.x, hypo.y, [hypo.z]
     hc = Point(clon, clat, cdeps[0])
     if step == 1:  # regular case, return full ruptures
-        surfin = src.get_planin(mags, nplanes)
-        surfaces = build_planar_surfaces(surfin, clon, clat, cdeps, shift_hypo)
+        planin = src.get_planin(mags, nplanes)
+        surfaces = build_planar_surfaces(planin, clon, clat, cdeps, shift_hypo)
         for m, mag in enumerate(mags):
             for n, np in enumerate(nplanes):
                 for d, hc_prob in enumerate(hc_probs):
@@ -223,14 +224,14 @@ class PointSource(ParametricSeismicSource):
 
     def get_planin(self, mags, nplanes):
         """
-        :return: array of dtype surfin_dt of shape (num_mags, num_planes)
+        :return: array of dtype planin_dt of shape (num_mags, num_planes)
         """
         msr = self.magnitude_scaling_relationship
-        surfin = numpy.zeros((len(mags), len(nplanes)), surfin_dt).view(
+        planin = numpy.zeros((len(mags), len(nplanes)), planin_dt).view(
             numpy.recarray)
         for m, mag in enumerate(mags):
             for n, np in enumerate(nplanes):
-                rec = surfin[m, n]
+                rec = planin[m, n]
                 rec['usd'] = self.upper_seismogenic_depth
                 rec['lsd'] = self.lower_seismogenic_depth
                 rec['rar'] = self.rupture_aspect_ratio
@@ -240,7 +241,7 @@ class PointSource(ParametricSeismicSource):
                 rec['dip'] = np.dip
                 rec['rake'] = np.rake
                 rec['dims'] = _get_rupture_dimensions(rec)
-        return surfin
+        return planin
 
     def _get_max_rupture_projection_radius(self, mag=None):
         """
@@ -254,8 +255,8 @@ class PointSource(ParametricSeismicSource):
             mag, _rate = self.get_annual_occurrence_rates()[-1]
         radius = []
         _, nplanes = zip(*self.nodal_plane_distribution.data)
-        for surfin in self.get_planin([mag], nplanes)[0]:
-            rup_length, rup_width, _ = _get_rupture_dimensions(surfin)
+        for planin in self.get_planin([mag], nplanes)[0]:
+            rup_length, rup_width, _ = _get_rupture_dimensions(planin)
             # the projection radius is half of the rupture diagonal
             radius.append(math.sqrt(rup_length ** 2 + rup_width ** 2) / 2.0)
         self.radius = max(radius)
@@ -265,9 +266,9 @@ class PointSource(ParametricSeismicSource):
         """
         :returns: half of maximum rupture's diagonal surface projection
         """
-        [[surfin]] = self.get_planin(
+        [[planin]] = self.get_planin(
             [rup.mag], [NodalPlane(rup.surface.strike, dip, rup.rake)])
-        rup_length, rup_width, _ = _get_rupture_dimensions(surfin)
+        rup_length, rup_width, _ = _get_rupture_dimensions(planin)
         return math.sqrt(rup_length ** 2 + rup_width ** 2) / 2.0
 
     def iter_ruptures(self, **kwargs):
@@ -389,9 +390,9 @@ class CollapsedPointSource(PointSource):
         """
         if mag is None:
             mag, _rate = self.get_annual_occurrence_rates()[-1]
-        [[surfin]] = self.get_planin(
+        [[planin]] = self.get_planin(
             [mag], [NodalPlane(self.strike, self.dip, self.rake)])
-        rup_length, rup_width, _ = _get_rupture_dimensions(surfin)
+        rup_length, rup_width, _ = _get_rupture_dimensions(planin)
         # the projection radius is half of the rupture diagonal
         self.radius = math.sqrt(rup_length ** 2 + rup_width ** 2) / 2.0
         return self.radius
